Wiring Up UI Elements In Employee Management Menu
I have made some scripts that i attached to the employeeInfoCards to
just hold the references for the name, gender etc. This will make it
easier to update them with employee information. I also did the same for
the close up view of a selected employee. The idea is by adding the
references on their own script, it will avoid confusion because there
are similar names for text elements being used in multiple places.