Implemented Firing Currently Hired Employees
Now with the press of the fire button, the employee is fired. This
required recoding the fire method but it was a long time coming. I also
improved to general behaviour of employees being selected on the page.
The first employee is always selected when you switch to the page and
more.