Designed General Outline Of Upgrades Menu
Just the basic outline of what the menu will do. I want to have a basic
concept for the rest of the menus first in case I come up will new
gameplay features along the way eg. the new Research mechanic. Just to
play it safe to avoid redoing a menu.