Major Work On Building Design
I completely changed the design of the business building. I have created
a room that can be customized. I have made a level of the building made
up of 4 rooms. When upgrading the building, after all rooms on a floor
are bought, a new level appears but the other rooms are locked until
bought.